SISRP-38681 - Show Committee Card in Adviser Dashboard - UX Development
